PP-289: Unique Jobid's upto 1 trillion

+++ src/cmds/scripts/pbs_db_schema.sql
@@ -73,14 +73,14 @@ CREATE SEQUENCE pbs.svr_id_seq;
  * Table pbs.server holds server instance information
  */
 CREATE TABLE pbs.server (
-    sv_name		TEXT		NOT NULL, /* the server id value */
+    sv_name			TEXT		NOT NULL, /* the server id value */
     sv_hostname		TEXT		NOT NULL, /* the actual server hostname - can change */
     sv_numjobs		INTEGER		NOT NULL,
-    sv_numque		INTEGER		NOT NULL,
-    sv_jobidnumber	INTEGER		NOT NULL,
+    sv_numque			INTEGER		NOT NULL,
+    sv_jobidnumber		BIGINT		NOT NULL,
     sv_svraddr		BIGINT		NOT NULL,
     sv_svrport		INTEGER		NOT NULL,
-    sv_savetm		TIMESTAMP	NOT NULL,
+    sv_savetm			TIMESTAMP	NOT NULL,
     sv_creattm		TIMESTAMP	NOT NULL,
     CONSTRAINT server_pk PRIMARY KEY (sv_name)
 );
